[IA-3443] OUCRC create - WIP
