+++ hat/assets/js/apps/Iaso/domains/orgUnits/configuration/OrgUnitChangeRequestConfigs.tsx
@@ -1,6 +1,6 @@
 import { makeStyles } from '@mui/styles';
-import { AddButton, commonStyles, useSafeIntl } from 'bluesquare-components';
-import React, { FunctionComponent } from 'react';
+import { commonStyles, useSafeIntl } from 'bluesquare-components';
+import React, { FunctionComponent, useState } from 'react';
 import { Box } from '@mui/material';
 import { useParamsObject } from '../../../routing/hooks/useParamsObject';
 import { baseUrls } from '../../../constants/urls';
@@ -10,7 +10,8 @@ import TopBar from '../../../components/nav/TopBarComponent';
 import MESSAGES from './messages';
 import { OrgUnitChangeRequestConfigsFilter } from './Filter/OrgUnitChangeRequestConfigsFilter';
 import { OrgUnitChangeRequestConfigsTable } from './Tables/OrgUnitChangeRequestConfigsTable';
-import { OrgUnitChangeRequestConfigDialog } from './Dialog/OrgUnitChangeRequestConfigDialog';
+import { OrgUnitChangeRequestConfigDialogCreate } from './Dialog/OrgUnitChangeRequestConfigDialogCreate';
+import { OrgUnitChangeRequestConfigDialogUpdate } from './Dialog/OrgUnitChangeRequestConfigDialogUpdate';
 
 const useStyles = makeStyles(theme => ({
     ...commonStyles(theme),
@@ -21,9 +22,13 @@ export const OrgUnitChangeRequestConfigs: FunctionComponent = () => {
         baseUrls.orgUnitsChangeRequestConfiguration,
     ) as unknown as OrgUnitChangeRequestConfigsParams;
     const { data, isFetching } = useGetOrgUnitChangeRequestConfigs(params);
+    const [isUpdateDialogOpen, setIsUpdateDialogOpen] =
+        useState<boolean>(false);
 
     const classes: Record<string, string> = useStyles();
     const { formatMessage } = useSafeIntl();
+
+    console.log("*** state isDialogOpen = ", isUpdateDialogOpen);
 
     return (
         <div>
@@ -35,11 +40,15 @@ export const OrgUnitChangeRequestConfigs: FunctionComponent = () => {
             <Box className={classes.containerFullHeightNoTabPadded}>
                 <OrgUnitChangeRequestConfigsFilter params={params} />
                 <Box mb={2} display="flex" justifyContent="flex-end">
-                    <OrgUnitChangeRequestConfigDialog
-                        titleMessage={formatMessage(MESSAGES.oucrcCreateModalTitle)}
-                        iconProps={{ disabled: false }}
-                        dataTestId="add-org-unit-config-button"
+
+                    <OrgUnitChangeRequestConfigDialogCreate
+                        iconProps={{}}
+                        openUpdateDialog={() => {
+                            console.log("*** openUpdateDialog ***");
+                            setIsUpdateDialogOpen(true);
+                        }}
                     />
+                    <OrgUnitChangeRequestConfigDialogUpdate isOpen={isUpdateDialogOpen} closeDialog={() => { setIsUpdateDialogOpen(false) }} />
                 </Box>
 
                 <OrgUnitChangeRequestConfigsTable
